requirements.txt
# 데이터 처리
pandas>=2.1.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# AI 분석
anthropic>=0.25.0

# 데이터 검증 모델
pydantic>=2.5.0

# 데이터베이스
sqlalchemy>=2.0.0

# 이상값 탐지
scipy>=1.11.0
scikit-learn>=1.3.0

# 보고서 생성
jinja2>=3.1.0
weasyprint>=60.0

# CLI
click>=8.1.0
rich>=13.7.0

# 환경변수
python-dotenv>=1.0.0

# 날짜 처리
python-dateutil>=2.8.0
